feat: add provider selection when initializing ecologits
