+++ ecologits/__init__.py
@@ -1,4 +1,4 @@
-from .ecologits import EcoLogits
+from ._ecologits import EcoLogits
 
 __version__ = "0.3.2"
 __all__ = [
