--- ecologits/ecologits.py
@@ -1,112 +1,0 @@
-import importlib.util
-
-from packaging.version import Version
-
-
-class EcoLogits:
-    """
-    EcoLogits instrumentor to initialize function patching for each provider.
-
-    By default, the initialization will be done on all available and compatible providers that are supported by the
-    library.
-
-    Examples:
-        EcoLogits initialization example with OpenAI.
-        ```python
-        from ecologits import EcoLogits
-        from openai import OpenAI
-
-        EcoLogits.init()
-
-        client = OpenAI(api_key="<OPENAI_API_KEY>")
-        response = client.chat.completions.create(
-            model="gpt-3.5-turbo",
-            messages=[
-                {"role": "user", "content": "Tell me a funny joke!"}
-            ]
-        )
-
-        # Get estimated environmental impacts of the inference
-        print(f"Energy consumption: {response.impacts.energy.value} kWh")
-        print(f"GHG emissions: {response.impacts.gwp.value} kgCO2eq")
-        ```
-
-    """
-
-    initialized = False
-
-    @staticmethod
-    def init() -> None:
-        """Initialization static method."""
-        if not EcoLogits.initialized:
-            init_instruments()
-            EcoLogits.initialized = True
-
-
-def init_instruments() -> None:
-    init_openai_instrumentor()
-    init_anthropic_instrumentor()
-    init_mistralai_instrumentor()
-    init_huggingface_instrumentor()
-    init_cohere_instrumentor()
-    init_google_instrumentor()
-    init_litellm_instrumentor()
-
-
-def init_openai_instrumentor() -> None:
-    if importlib.util.find_spec("openai") is not None:
-        from ecologits.tracers.openai_tracer import OpenAIInstrumentor
-
-        instrumentor = OpenAIInstrumentor()
-        instrumentor.instrument()
-
-
-def init_anthropic_instrumentor() -> None:
-    if importlib.util.find_spec("anthropic") is not None:
-        from ecologits.tracers.anthropic_tracer import AnthropicInstrumentor
-
-        instrumentor = AnthropicInstrumentor()
-        instrumentor.instrument()
-
-
-def init_mistralai_instrumentor() -> None:
-    if importlib.util.find_spec("mistralai") is not None:
-        from ecologits.tracers.mistralai_tracer import MistralAIInstrumentor
-
-        instrumentor = MistralAIInstrumentor()
-        instrumentor.instrument()
-
-
-def init_huggingface_instrumentor() -> None:
-    if importlib.util.find_spec("huggingface_hub") is not None:
-        from huggingface_hub import __version__
-
-        if Version(__version__) >= Version("0.22.0"):
-            from ecologits.tracers.huggingface_tracer import HuggingfaceInstrumentor
-
-            instrumentor = HuggingfaceInstrumentor()
-            instrumentor.instrument()
-
-
-def init_cohere_instrumentor() -> None:
-    if importlib.util.find_spec("cohere") is not None:
-        from ecologits.tracers.cohere_tracer import CohereInstrumentor
-
-        instrumentor = CohereInstrumentor()
-        instrumentor.instrument()
-
-
-def init_google_instrumentor() -> None:
-    if importlib.util.find_spec("google.generativeai") is not None:
-        from ecologits.tracers.google_tracer import GoogleInstrumentor
-
-        instrumentor = GoogleInstrumentor()
-        instrumentor.instrument()
-
-
-def init_litellm_instrumentor() -> None:
-    if importlib.util.find_spec("litellm") is not None:
-        from ecologits.tracers.litellm_tracer import LiteLLMInstrumentor
-
-        instrumentor = LiteLLMInstrumentor()
-        instrumentor.instrument()
